changed some std::int64_t to std::uint8_t

+++ src/Grid15/GridHelp.h
@@ -34,8 +34,8 @@ namespace Grid15
  */
 namespace GridHelp
 {
-bool validMove(const std::int64_t tileNum, const Grid &grid);
-bool validMove(const std::int64_t tileX, const std::int64_t tileY, const Grid &grid);
+bool validMove(const std::uint8_t tileNum, const Grid &grid);
+bool validMove(const std::uint8_t tileX, const std::uint8_t tileY, const Grid &grid);
 bool hasWon(const Grid &grid);
 
 void copyGridArray(const Grid::gridArray_t &newGrid, Grid &grid);
